backend/requirements.txt
python-dotenv
fastapi
uvicorn
gunicorn
celery
redis
psycopg2-binary #개발용
# psycopg2 #배포용
sqlalchemy
pytest
jinja2
python-multipart
opencv-python
keybert
sentence-transformers
spacy
bert_score
evaluate
rouge_score
